requirements.txt
python-dotenv
pandas
requests
numpy
